feat(markdown-panel): add autosize textarea
