+++ frontend/package.json
@@ -33,6 +33,7 @@
     "react-markdown": "^10.1.0",
     "react-rating": "^2.0.5",
     "react-router-dom": "^7.9.5",
+    "react-textarea-autosize": "^8.5.9",
     "rehype-raw": "^7.0.0",
     "rehype-sanitize": "^6.0.0",
     "tailwindcss": "^4.1.18",
